+++ test/unit/compiler/passes/report-duplicate-labels.spec.js
@@ -12,12 +12,20 @@ describe("compiler pass |reportDuplicateLabels|", function() {
   describe("in a sequence", function() {
     it("reports labels duplicate with labels of preceding elements", function() {
       expect(pass).to.reportError("start = a:'a' a:'a'", {
-        message: "Label \"a\" is already defined at line 1, column 9.",
+        message: "Label \"a\" is already defined",
         location: {
           source: undefined,
           start: { offset: 14, line: 1, column: 15 },
-          end: { offset: 19, line: 1, column: 20 }
-        }
+          end: { offset: 15, line: 1, column: 16 }
+        },
+        diagnostics: [{
+          message: "Original label location",
+          location: {
+            source: undefined,
+            start: { offset: 8, line: 1, column: 9 },
+            end: { offset: 9, line: 1, column: 10 }
+          }
+        }]
       });
     });
 
@@ -45,12 +53,20 @@ describe("compiler pass |reportDuplicateLabels|", function() {
   describe("in outer sequence", function() {
     it("reports labels duplicate with labels of preceding elements", function() {
       expect(pass).to.reportError("start = a:'a' (a:'a')", {
-        message: "Label \"a\" is already defined at line 1, column 9.",
+        message: "Label \"a\" is already defined",
         location: {
           source: undefined,
           start: { offset: 15, line: 1, column: 16 },
-          end: { offset: 20, line: 1, column: 21 }
-        }
+          end: { offset: 16, line: 1, column: 17 }
+        },
+        diagnostics: [{
+          message: "Original label location",
+          location: {
+            source: undefined,
+            start: { offset: 8, line: 1, column: 9 },
+            end: { offset: 9, line: 1, column: 10 }
+          }
+        }]
       });
     });